+++ bittensor/neuron.py
@@ -18,38 +18,7 @@
 import argparse
 from munch import Munch
 from loguru import logger
-
-import multiprocessing.managers
-from multiprocessing import Process, Manager
-from multiprocessing.managers import BaseManager, NamespaceProxy, BaseProxy, AutoProxy
-
 import bittensor
-
-def AutoProxy(token, serializer, manager=None, authkey=None,
-              exposed=None, incref=True, manager_owned=False):
-    '''
-    Return an auto-proxy for `token`
-    '''
-    _Client = multiprocessing.managers.listener_client[serializer][1]
-
-    if exposed is None:
-        conn = _Client(token.address, authkey=authkey)
-        try:
-            exposed = dispatch(conn, None, 'get_methods', (token,))
-        finally:
-            conn.close()
-
-    if authkey is None and manager is not None:
-        authkey = manager._authkey
-    if authkey is None:
-        authkey = multiprocessing.process.current_process().authkey
-
-    ProxyType = multiprocessing.managers.MakeProxyType('AutoProxy[%s]' % token.typeid, exposed)
-    proxy = ProxyType(token, serializer, manager=manager, authkey=authkey,
-                      incref=incref, manager_owned=manager_owned)
-    proxy._isauto = True
-    return proxy
-multiprocessing.managers.AutoProxy = AutoProxy
 
 class Neuron:
     r"""
