+++ tests/unit_tests/bittensor_tests/test_forward.py
@@ -5,8 +5,8 @@ from unittest.mock import MagicMock
 from torch.autograd import Variable
 
 bittensor.init()
-bittensor.dendrite.forward = MagicMock(return_value = [torch.tensor([]), [0], ['']]) 
-bittensor.dendrite.backward = MagicMock(return_value = [torch.tensor([]), [0], ['']]) 
+bittensor.dendrite._dendrite.forward = MagicMock(return_value = [torch.tensor([]), [0], ['']]) 
+bittensor.dendrite._dendrite.backward = MagicMock(return_value = [torch.tensor([]), [0], ['']]) 
 
 neuron_pb2 = bittensor.proto.Neuron(
     version = bittensor.__version__,
@@ -18,44 +18,44 @@ neuron_pb2 = bittensor.proto.Neuron(
 def test_dendrite_forward_tensor_shape_error():
     x = torch.rand(3, 3, 3)
     with pytest.raises(ValueError):
-        bittensor.forward_tensor( neurons=[neuron_pb2], inputs=[x])
+        bittensor.dendrite.forward_tensor( neurons=[neuron_pb2], inputs=[x])
 
 def test_dendrite_forward_image_shape_error():
     x = torch.rand(3, 3, 3)
     with pytest.raises(ValueError):
-        bittensor.forward_image( neurons=[neuron_pb2], inputs=[x])
+        bittensor.dendrite.forward_image( neurons=[neuron_pb2], inputs=[x])
 
 def test_dendrite_forward_text_shape_error():
     x = torch.rand(3, 3, 3)
     with pytest.raises(ValueError):
-        bittensor.forward_image( neurons=[neuron_pb2], inputs=[x])
+        bittensor.dendrite.forward_image( neurons=[neuron_pb2], inputs=[x])
 
 def test_dendrite_forward_text():
     x = torch.tensor([[1,2,3,4],[5,6,7,8]], dtype=torch.long)
-    bittensor.dendrite.forward = MagicMock(return_value = [ [torch.zeros([2, 4, bittensor.__network_dim__])], [0], ['']]) 
-    codes, tensors  = bittensor.forward_text( neurons=[neuron_pb2], inputs=[x])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [torch.zeros([2, 4, bittensor.__network_dim__])], [0], ['']]) 
+    codes, tensors  = bittensor.dendrite.forward_text( neurons=[neuron_pb2], inputs=[x])
     assert codes[0].item() == bittensor.proto.ReturnCode.Success
     assert list(tensors[0].shape) == [2, 4, bittensor.__network_dim__]
 
 def test_dendrite_forward_image():
     x = torch.tensor([ [ [ [ [ 1 ] ] ] ] ])
-    bittensor.dendrite.forward = MagicMock(return_value = [ [torch.zeros([1, 1, bittensor.__network_dim__])] , [0], ['']]) 
-    codes, tensors  = bittensor.forward_image( neurons=[neuron_pb2], inputs=[x])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [torch.zeros([1, 1, bittensor.__network_dim__])] , [0], ['']]) 
+    codes, tensors  = bittensor.dendrite.forward_image( neurons=[neuron_pb2], inputs=[x])
     assert codes[0].item() == bittensor.proto.ReturnCode.Success
     assert list(tensors[0].shape) == [1, 1, bittensor.__network_dim__]
 
 def test_dendrite_forward_tensor():
     x = torch.rand(3, 3, bittensor.__network_dim__)
-    bittensor.dendrite.forward = MagicMock(return_value = [ [torch.zeros([3, 3, bittensor.__network_dim__])], [0], ['']]) 
-    codes, tensors = bittensor.forward_tensor( neurons=[neuron_pb2], inputs=[x])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [torch.zeros([3, 3, bittensor.__network_dim__])], [0], ['']]) 
+    codes, tensors = bittensor.dendrite.forward_tensor( neurons=[neuron_pb2], inputs=[x])
     assert codes[0].item() == bittensor.proto.ReturnCode.Success
     assert list(tensors[0].shape) == [3, 3, bittensor.__network_dim__]
 
 def test_dendrite_forward_tensor_pass_through_text():
     x = torch.rand(3, 3)
     y = torch.zeros([3, 3, bittensor.__network_dim__])
-    bittensor.dendrite.forward = MagicMock(return_value = [ [y, y, y] , [0, 0, 0], ['','',''] ]) 
-    codes, tensors = bittensor.forward_text( neurons=[neuron_pb2, neuron_pb2, neuron_pb2], inputs=[x, x, x])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [y, y, y] , [0, 0, 0], ['','',''] ]) 
+    codes, tensors = bittensor.dendrite.forward_text( neurons=[neuron_pb2, neuron_pb2, neuron_pb2], inputs=[x, x, x])
     assert codes[0].item() == bittensor.proto.ReturnCode.Success
     assert codes[1].item() == bittensor.proto.ReturnCode.Success
     assert codes[2].item() == bittensor.proto.ReturnCode.Success
@@ -66,8 +66,8 @@ def test_dendrite_forward_tensor_pass_through_text():
 def test_dendrite_forward_tensor_pass_through_image():
     x = torch.rand(3, 3, 3, 3, 3)
     y = torch.zeros([3, 3, bittensor.__network_dim__])
-    bittensor.dendrite.forward = MagicMock(return_value = [ [y, y, y] , [0, 0, 0], ['','','']]) 
-    codes, tensors = bittensor.forward_image( neurons=[neuron_pb2, neuron_pb2, neuron_pb2], inputs=[x, x, x])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [y, y, y] , [0, 0, 0], ['','','']]) 
+    codes, tensors = bittensor.dendrite.forward_image( neurons=[neuron_pb2, neuron_pb2, neuron_pb2], inputs=[x, x, x])
     assert codes[0].item() == bittensor.proto.ReturnCode.Success
     assert codes[1].item() == bittensor.proto.ReturnCode.Success
     assert codes[2].item() == bittensor.proto.ReturnCode.Success
@@ -78,8 +78,8 @@ def test_dendrite_forward_tensor_pass_through_image():
 def test_dendrite_forward_tensor_pass_through_tensor():
     x = torch.rand(3, 3, bittensor.__network_dim__)
     y = torch.zeros([3, 3, bittensor.__network_dim__])
-    bittensor.dendrite.forward = MagicMock(return_value = [ [y, y, y] , [0, 0, 0], ['','','']]) 
-    codes, tensors = bittensor.forward_tensor( neurons=[neuron_pb2, neuron_pb2, neuron_pb2], inputs=[x, x, x])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [y, y, y] , [0, 0, 0], ['','','']]) 
+    codes, tensors = bittensor.dendrite.forward_tensor( neurons=[neuron_pb2, neuron_pb2, neuron_pb2], inputs=[x, x, x])
     assert codes[0].item() == bittensor.proto.ReturnCode.Success
     assert codes[1].item() == bittensor.proto.ReturnCode.Success
     assert codes[2].item() == bittensor.proto.ReturnCode.Success
@@ -90,18 +90,18 @@ def test_dendrite_forward_tensor_pass_through_tensor():
 def test_dendrite_backward():
     x = Variable(torch.ones(1, 1), requires_grad=True)
     y = torch.ones(1, 1)
-    bittensor.dendrite.forward = MagicMock(return_value = [ [y], [0], ['']]) 
-    bittensor.dendrite.backward = MagicMock(return_value = [ [y], [0], ['']]) 
-    codes, tensors = bittensor.forward_text( neurons=[ neuron_pb2 ], inputs=[ x ])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [y], [0], ['']]) 
+    bittensor.dendrite._dendrite.backward = MagicMock(return_value = [ [y], [0], ['']]) 
+    codes, tensors = bittensor.dendrite.forward_text( neurons=[ neuron_pb2 ], inputs=[ x ])
     tensors[0].backward()
     assert x.grad.item() == 1
 
 def test_dendrite_backward_large():
     x = Variable(torch.ones(1, 2), requires_grad=True)
     y = torch.ones(1, 2)
-    bittensor.dendrite.forward = MagicMock(return_value = [ [y], [0], ['']]) 
-    bittensor.dendrite.backward = MagicMock(return_value = [ [y], [0], ['']]) 
-    codes, tensors = bittensor.forward_text( neurons=[ neuron_pb2 ], inputs=[ x ])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [y], [0], ['']]) 
+    bittensor.dendrite._dendrite.backward = MagicMock(return_value = [ [y], [0], ['']]) 
+    codes, tensors = bittensor.dendrite.forward_text( neurons=[ neuron_pb2 ], inputs=[ x ])
     tensors[0].sum().backward()
     assert list(x.grad.shape) == [1, 2]
     assert x.grad.tolist() == y.tolist()
@@ -114,9 +114,9 @@ def test_dendrite_backward_multiple():
     y2 = torch.ones(1, 1)
     y3 = torch.ones(1, 1)
 
-    bittensor.dendrite.forward = MagicMock(return_value = [ [y1, y2, y3], [0,0,0], ['','','']]) 
-    bittensor.dendrite.backward = MagicMock(return_value = [ [y1, y2, y3], [0,0,0], ['','','']]) 
-    codes, tensors = bittensor.forward_text( neurons=[ neuron_pb2, neuron_pb2, neuron_pb2 ], inputs=[ x1, x2, x3 ])
+    bittensor.dendrite._dendrite.forward = MagicMock(return_value = [ [y1, y2, y3], [0,0,0], ['','','']]) 
+    bittensor.dendrite._dendrite.backward = MagicMock(return_value = [ [y1, y2, y3], [0,0,0], ['','','']]) 
+    codes, tensors = bittensor.dendrite.forward_text( neurons=[ neuron_pb2, neuron_pb2, neuron_pb2 ], inputs=[ x1, x2, x3 ])
     tensors[0].backward()
     assert list(x1.grad.shape) == [1, 1]
     assert x1.grad.tolist() == y1.tolist()
